added router for browser routing
