+++ src/Cart.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+export default function Cart() {
+    return (
+        <h1>Cart</h1>
+    );
+}
