+++ src/Detail.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+export default function Detail() {
+    return (
+        <h1>Detail</h1>
+    );
+}
